examples/requirements.txt
# TraceCraft Examples - All Dependencies
#
# This file lists all dependencies needed to run all examples.
# For minimal installations, see section-specific requirements below.

# Core (already included with tracecraft)
# tracecraft

# Framework Integrations
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
llama-index-core>=0.11.0
llama-index-llms-openai>=0.2.0
pydantic-ai>=0.1.0
openai>=1.50.0

# Evaluation Frameworks
deepeval>=1.0.0
ragas>=0.2.0
datasets>=3.0.0
mlflow>=2.16.0

# Alerting (HTTP requests)
httpx>=0.27.0

# Visualization (included with tracecraft)
# rich

# Optional: For async examples
# asyncio (standard library)

# Development
pytest>=8.0
pytest-asyncio>=0.24.0
